wip: move all core Koa code into a seperate module
This change helps make Koa more testable

refs #24
